// src/app/interfaces/book.interfaces.ts
// interfaces for the book object
export interface Book {
  id: number;
  title: string;
  author: string;
  description: string;
  publishedYear: number;
  price: number;
  imageUrl: string;
  genre: string;
  rating: number;
  pages: number;
  language: string;
  publisher: string;
  isbn: string;
}